requirements.txt
tensorflow==2.4.1
streamlit==0.80.0
numpy==1.19.5
pillow==8.1.0